Address Copilot review: break IHV circular import, clear SSE reconnect timers, reconcile audit counts
Move pass validations and getCellCompatibility into hardwarePassCompatibility.ts so the matrix no longer imports IHVIntegrationPanel. Clear reconnectTimeoutRef before every reconnect schedule and on connectSSE entry. Align Tech Debt audit summary with the current GitHub open/closed snapshot (22 closed / 33 open).
